Fix test error logs in EngagementForm tests (#2116)
* Fix test error logs in EngagementForm tests

* add savedEngagements to useEffect

* use factory mock data

* Remove unused import
