+++ met-web/src/components/engagement/form/EngagementFormTabs/EngagementTabsContext.tsx
@@ -197,8 +197,10 @@ export const EngagementTabsContextProvider = ({ children }: { children: React.Re
     };
 
     useEffect(() => {
-        loadTeamMembers();
-    }, []);
+        if (savedEngagement.id) {
+            loadTeamMembers();
+        }
+    }, [savedEngagement]);
 
     const [settings, setSettings] = useState<EngagementSettings>(createDefaultEngagementSettings());
     const [settingsLoading, setSettingsLoading] = useState(true);
